requirements.txt
ultralytics==8.3.229
streamlit==1.29.0
pillow==10.1.0
numpy==1.25.0